test: Merge file and camp tests
